+++ setup.py
@@ -63,6 +63,6 @@ setup(
         "Slack": "https://join.slack.com/t/nuclia-community/shared_invite/zt-2ldfznncw-LW6GJjOicdEA18kgI~95Xw",
     },
     extras_require={
-        "langchain": ["langchain_core>=0.3.29"],
+        "ai": ["litellm"],
     },
 )
